pyproject.toml
[project]
name = "beavr-bench"
version = "1.0.0"
description = "BEAVR Bench: Simulation benchmark suite for testing and evaluating physical AI algorithms"
authors = [
    { name = "BEAVR Team", email = "team@beavr.ai" },
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10,<4.0"
keywords = ["robotics", "deep learning", "pytorch", "teleoperation", "vr", "mujoco", "isaac-lab", "lerobot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "pyzmq",
    "draccus",
    "numpy",
    "mjlab>=0.1.0",
    "mujoco-warp",
    "aiohttp>=3.13.2",
    "lerobot @ git+https://github.com/huggingface/lerobot.git",
    "torch",
    "torchvision",
]

[project.scripts]
beavr-eval = "beavr_bench.scripts.eval:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/beavr_bench"]

[tool.ruff]
target-version = "py310"
line-length = 110
exclude = ["tests/artifacts/**/*.safetensors", "*_pb2.py", "*_pb2_grpc.py"]

[tool.ruff.lint]
# E, W: pycodestyle errors and warnings
# F: PyFlakes
# I: isort
# UP: pyupgrade
# B: flake8-bugbear (good practices, potential bugs)
# C4: flake8-comprehensions (more concise comprehensions)
# A: flake8-builtins (shadowing builtins)
# SIM: flake8-simplify
# RUF: Ruff-specific rules
# D: pydocstyle (for docstring style/formatting)
# S: flake8-bandit (some security checks, complements Bandit)
# T20: flake8-print (discourage print statements in production code)
# N: pep8-naming
# TODO: Uncomment rules when ready to use
select = [
    "E", "W", "F", "I", "B", "C4", "T20", "N", "UP" # "SIM", "A", "S", "D", "RUF"
]
ignore = [
    "E501", # Line too long
    "T201", # Print statement found
    "T203", # Pprint statement found
    "B008", # Perform function call in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["beavr"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
